Minor fixes
Sandcastle
- Fixed some bugs in the AddNamespaceGroups tool and made inclusion of
the root container more consistent based on the content and whether or
not a root namespace container is present.  Closes #30.

Sandcastle Help File Builder
- Fixed a few theme and font size issues in the VSPackage.
- Added code to check for the shadow copy folder when determining the
tool location in the VSPackage.  Fixes #31.
- Added Microsoft.VisualStudio.TestTools.UITest.Playback to the default
list of "ignore if unresolved" assemblies in the binding redirection
plug-in default configuration.

Presentation Styles
- Fixed the keyword search script in the VS2013 presentation style so
that it correctly decodes the file indices when there are more than
65,535 files.  Fixes #28.