Bug fixes and a couple of new features
Sandcastle
- Fixed the extension method parameter comparisons in the MRefBuilder
extension method add-in so that it doesn't add extension methods to
types with matching method signatures.  Fixes #119.
- Added support for documenting public compiler generated types and
members.  Closes #106.

Sandcastle Help File Builder
- Added support for documenting public compiler generated types and
members.  Closes #106.

Presentation Styles
- Added an includeEnumValues transformation argument to all presentation
styles.  If set to true, the default, a column for the numeric value of
each field is included in enumerated type topics.  If set to false, the
numeric value column is excluded.  Closes #120.