Bug fixes, and added Reflection Data Manager
General
- Removed the scripts and configuration files used to build the
reflection data.  The new Reflection Data Manager tool is now used to do
that.  Eventually, it will support a command line interface for use in
the master build script but that will come later.

Sandcastle
- Fixed the API filter in MRefBuilder so that an internal type in one
assembly doesn't hide a public type of the same name in another assembly
if the internal one is seen first.
- Added the Reflection Data Manager tool which is used to define
reflection data set files and build the reflection data for the various
frameworks.
- Removed the old Prototype style and Help 2 production XSL
transformations.

Sandcastle Help File Builder
- Fixed the HelpLibraryManagerLauncher tool so that it adds the default
catalog name to the command line arguments if one is not explicitly
specified.  Fixes #122.

Presentation Styles
- Added the missing CodeExample.png image to the Open XML presentation
style.  Fixes #129.
- Updated the order of the sections on the member list pages in all
presentation styles to more closely match the order on MSDN.  Also
updated the TOC transform to order the type subtopics so that they're
consistent with the order in the member list pages.