fix: Assemblies enumerated multiple times
Whenever you select "Help File", "Components", "Plug-Ins" or "Transform Args" project property tabs the UI freezes and there is a noticable delay in loading the tab.
It appears the folders are traversed over and over again, the files are loaded over and over again.
The traversed paths are not recorded, which causes them to be traversed again whenever the ComponentUtilities.CreateComponentContainer(...) method is called.

The fix addressed the issue by making catalog and searched folders objects static.