Improved property page responsiveness
Added a component cache that loads the component information
asynchronously and is shared amongst the property pages that need it to
improve responsiveness.  Changing the component path property also
triggers a refresh which should remove the need to unload and reload a
project after changing it to see the components in the new location.