Miscellaneous Updates
- Modified the MPF and SHFB project system to allow for excluding the
content of the output path and working
folders when doing Show All Files.  This is a quick hack to work around
the very slow performance when using
Show All Files due to the large number of files in the output and
working folders when website output is
produced.  The output folder itself is still included but subfolders are
not.  This would prevent the inclusion
of subfolders in the project in cases where the project does not produce
website output but chances are most
people won't want to include these folders in the project anyway.  Fixes
#198.
- Changed the Lightweight website style plug-in to generate the TOC info
files in parallel to improve performance.
- Changed the website full text index generation so that it processes
the files in parallel to improve performance.