Build engine fix
Modified the build engine so that the help project being built attempts
to retrieve the SolutionPath macro from the help project in order to use
it as the default for the solution macros when extracting the targets
from project documentation sources.  This allows documentation source
projects that use solution macros in their output path to resolve
properly as long as the help project is built as part of the containing
solution.  Fixes #238.