Various fixes and updates
Documentation
- Updated the Files and Folders topic to explain linked items and how
they are used in a project.  Closes #319.

Sandcastle
- Fixed MRefBuilder so that it resolves a missing mscorlib with a major
version number 255 to System.Runtime which is common for frameworks like
.NETCore that redirect all of their system types there.
- Added support for the the .NETCore and .NETPortable Frameworks
distributed as NuGet packages (effective with VS2015 Update 3).  Closes
#322.
- Added support for the .NET Framework 4.6.2.  Closes #325.

Sandcastle Help File Builder
- Fixed up case sensitivity issues and issues with environment variables
that match property names when loading project properties.  Fixes #267.
Fixes #321.
- Fixed Multi-Format Output Component so that it recreates the document
for each format from the XML rather than cloning it to workaround an
occassional issue with the clone that results in an "invalid parent"
error.  Fixes #288.
- Added code to bypass the assembly binding redirection plug-in
execution and issue a warning if the MRefBuilder configuration file is
not found (i.e. the Additional Content Only plug-in is enabled).  Closes
#314.
- Cleared out some old TODO items.

Presentation Styles
- Fixed all presenation styles so that they generate the root group
topic correctly when namespace grouping is used without a root namespace
container.  Fixes #316.
- Updated the Open XML build task so that it includes the `align` and
`posOffset` elements as parent elements that can contain text that does
not need to be wrapped in a run element.  Fixes #320.