Various updates and fixes
General
- Fixed the guided installer so that it handles "Program Files" folder
references properly on 32-bit systems.  Fixes #450.
- Updated the MAML IntelliSense page to install the schema reference so
that it uses the SHFB copy of the files.

Documentation
- Added a lineBreak element topic to the MAML Guide.
- Revised the topic on how to install the MAML schemas so that it shows
how to use the SHFB copies.

Sandcastle
- Fixed MRefBuilder so that it doesn't match a nested type to an
incorrect file as it's source code location under certain circumstances.
- Updated the code block component so that it removes the trailing "REM"
on regions imported from batch scripts.  Fixes #428.

Sandcastle Help File Builder
- Modified the build engine so that it compares the full strong name
when determining whether or not to exclude reference assemblies that are
in the selected reflection data set.  This allows for like-named
assemblies with a different version to be included as references when
generating reflection information (i.e. NuGet versions of the .NETCore
and .NETStandard framework assemblies).

Presentation Styles
- Fixed the XSL transformations in the Open XML and Markdown
presentation styles to prevent insertion of extraneous spaces when
normalizing text.
- Fixed the script in the VS2010 and VS2013 presentation styles so that
it only inserts a trailing space after LST elements when really needed.
Fixes #451.
- Fixed the markdown content build task so that it handles duplicate
in-page link target addresses and issues a warning if any are found.
- Added support for the MAML lineBreak element.  This inserts a line
break in the text without creating a new paragraph.  Closes #431.