Minor fixes and updates
General
- Updated the Sandcastle Installer to note the MSBuild 2015 Tools
installation requirement for use with the standalone GUI.  Closes #467.

Sandcastle
- Fixed MRefBuilder so that if an overridden member is removed, the
matching overridden/virtual members in the base types are also removed
to prevent them showing up in place of the removed override.  Fixes
#500.
- Modified AddFilenames.xsl and added a more specific exception message
in XslTransform to clarify the solution to issues caused by the FIPS
policy being enabled.  Closes #498.

Sandcastle Help File Builder
- Fixed the Help Library Manager Launcher tool so that it finds the
correct registry keys when SSMS is installed which can result in the
64-bit parent key existing without the expected subkeys.  Fixes #501.