Changes to workaround a NuGet issue
Modified the names and handling of the relationship files in the Open
XML presentation style to work around an issue with the latest version
of NuGet which adds a duplicate entry to its content types file for the
".rels" file type which can break third-party tools that parse the
package.  Fixes #555.