Various updates and fixes
Changes made in this release:

General
- Added support for colorizing Xamarin Android code blocks (.axml).
Closes #584.
- Added a reflection data set file for the .NET Framework 4.7.1.

Sandcastle
- MRefBuilder: Merged change from Emmanuel Osimosu to fix a null
reference exception.  Fixes #580.
- BuildAssembler: Updated the Shared Content Component so that it looks
for an "undefined" attribute on content items.  If the content item is
not found, it uses the value of the "undefined" attribute as the content
item value.  This allows unknown code languages to be added without
having to modify all of the presentation styles.
- BuildAssembler: Fixed the Syntax Component so that if an unknown
language ID is encountered, it attempts to find a friendly name using
the code colorizer definitions or, failing that, uses the language ID as
the title rather than omitting it completely.  This allows unknown
languages to be added without the need for modifying the presentation
styles or code colorizer rules.

Sandcastle Help File Builder
- Fixed the project platform check so that it works properly with
.NETPortable project types.  Fixes #566.
- Fixed the SHFB package so that it doesn't throw an exception when
devenv.exe is invoked for a command line build.  Fixes #577.
- Fixed the Additional Reference Links plug-in so that it doesn't fail
if the "OutDir" variable is undefined.  Fixes the problem referenced in
#586.
- Added XML comments validation checks to throw out those with
ill-formed content.  Fixes #560, #569, and #574.
- Added another member ID fix-up to the Member ID Fix-Ups plug-in
submitted by Andy Dennie.  Closes #588.

Presentation Styles
- Updated all presentation styles to add the "undefined" element to code
language content item elements to pass a value to use for the language
if a named shared content item doesn't exist for it.