NuGet SHFB package property file change
Changed the NuGet SHFB package build properties file to use `BuildingInsideVisualStudio` instead of an empty SHFBROOT value to determine when to define SHFBROOT for command line builds.  Fixes #955 and should also resolve #920 but only in cases where projects use a package version containing this change (a version later the 2022.10.15.1).  See the workaround in #955 for earlier packages.